refactor: move geography req to folders
