+++ internal/router/geography.go
@@ -2,10 +2,11 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	carryHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	geographyHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/geography"
 )
 
-func MountGeographyRoutes(api chi.Router, hdGeography *handler.GeographyHandler, hdCarry *handler.CarryHandler) {
+func MountGeographyRoutes(api chi.Router, hdGeography *geographyHandler.GeographyHandler, hdCarry *carryHandler.CarryHandler) {
 	api.Route("/localities", func(r chi.Router) {
 		r.Post("/", hdGeography.Create)
 		r.Get("/reportSellers", hdGeography.CountSellersByLocality)
