+++ internal/router/router.go
@@ -6,6 +6,7 @@ import (
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
 	buyerHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/buyer"
 	empHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee"
+	geographyHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/geography"
 	inbHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/inbound_order"
 	productHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/product"
 	ProductRecordHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/product_record"
@@ -22,7 +23,7 @@ func NewAPIRouter(
 	hdProduct *productHandler.ProductHandler,
 	hdProductBatches *handler.ProductBatchesHandler,
 	hdPurchaseOrder *handler.PurchaseOrderHandler,
-	hdGeography *handler.GeographyHandler,
+	hdGeography *geographyHandler.GeographyHandler,
 	hdInboundOrder *inbHandler.InboundOrderHandler,
 	hdCarry *handler.CarryHandler,
 	hdProductRecord *ProductRecordHandler.ProductRecordHandler,
